requirements.txt
asyncpg
httpx
beautifulsoup4
fastapi
uvicorn
python-dotenv
Pillow
aiofiles